[agenda] Split up large views file
