+++ src/pretalx/agenda/__init__.py
@@ -6,3 +6,4 @@ class AgendaConfig(AppConfig):
 
     def ready(self):
         from .phrases import AgendaPhrases  # noqa
+        from . import permissions  # noqa
